Added user auth.

+++ src/resolvers/todo.go
@@ -1,6 +1,9 @@
 package resolvers
 
 import (
+	"fmt"
+	"time"
+
 	"github.com/go-pg/pg"
 	"github.com/graphql-go/graphql"
 
@@ -9,13 +12,34 @@ import (
 
 func GetTodos(params graphql.ResolveParams) (interface{}, error) {
 	database := params.Context.Value("database").(*pg.DB)
+	user := params.Context.Value("user").(*structs.Claims)
+
+	if !user.VerifyExpiresAt(time.Now().Unix(), true) {
+		return nil, fmt.Errorf("Not authenticated.")
+	}
 
 	skip, _ := params.Args["skip"].(int)
 	limit, _ := params.Args["limit"].(int)
+	filter, _ := params.Args["filter"].(string)
 
 	var todos []structs.Todo
+	query := database.
+		Model(&todos).
+		Where("user_id = ?", user.ID)
 
-	if err := database.Model(&todos).OrderExpr("id DESC").Offset(skip).Limit(limit).Select(); err != nil {
+	switch filter {
+	case "active":
+		query = query.Where("completed IS NOT TRUE")
+		break
+	case "complete":
+		query = query.Where("completed IS TRUE")
+	}
+
+	if err := query.
+		OrderExpr("id DESC").
+		Offset(skip).
+		Limit(limit).
+		Select(); err != nil {
 		return nil, err
 	}
 
@@ -24,8 +48,17 @@ func GetTodos(params graphql.ResolveParams) (interface{}, error) {
 
 func TodosLeft(params graphql.ResolveParams) (interface{}, error) {
 	database := params.Context.Value("database").(*pg.DB)
+	user := params.Context.Value("user").(*structs.Claims)
 
-	count, err := database.Model(&structs.Todo{}).Where("completed IS NOT TRUE").Count()
+	if !user.VerifyExpiresAt(time.Now().Unix(), true) {
+		return nil, fmt.Errorf("Not authenticated.")
+	}
+
+	count, err := database.
+		Model(&structs.Todo{}).
+		Where("user_id = ?", user.ID).
+		Where("completed IS NOT TRUE").
+		Count()
 
 	if err != nil {
 		return nil, err
@@ -36,12 +69,18 @@ func TodosLeft(params graphql.ResolveParams) (interface{}, error) {
 
 func CreateTodo(params graphql.ResolveParams) (interface{}, error) {
 	database := params.Context.Value("database").(*pg.DB)
+	user := params.Context.Value("user").(*structs.Claims)
+
+	if !user.VerifyExpiresAt(time.Now().Unix(), true) {
+		return nil, fmt.Errorf("Not authenticated.")
+	}
 
 	text, _ := params.Args["text"].(string)
 
 	newTodo := structs.Todo{
 		Text:      text,
 		Completed: false,
+		UserID:    user.ID,
 	}
 
 	if err := database.Insert(&newTodo); err != nil {
@@ -53,6 +92,11 @@ func CreateTodo(params graphql.ResolveParams) (interface{}, error) {
 
 func MarkTodoCompleted(params graphql.ResolveParams) (interface{}, error) {
 	database := params.Context.Value("database").(*pg.DB)
+	user := params.Context.Value("user").(*structs.Claims)
+
+	if !user.VerifyExpiresAt(time.Now().Unix(), true) {
+		return nil, fmt.Errorf("Not authenticated.")
+	}
 
 	id, _ := params.Args["id"].(int)
 	completed, _ := params.Args["completed"].(bool)
@@ -63,6 +107,10 @@ func MarkTodoCompleted(params graphql.ResolveParams) (interface{}, error) {
 
 	if err := database.Select(&toUpdate); err != nil {
 		return nil, err
+	}
+
+	if toUpdate.UserID != user.ID {
+		return nil, fmt.Errorf("Not authrorized.")
 	}
 
 	toUpdate.Completed = completed
@@ -76,6 +124,11 @@ func MarkTodoCompleted(params graphql.ResolveParams) (interface{}, error) {
 
 func EditTodo(params graphql.ResolveParams) (interface{}, error) {
 	database := params.Context.Value("database").(*pg.DB)
+	user := params.Context.Value("user").(*structs.Claims)
+
+	if !user.VerifyExpiresAt(time.Now().Unix(), true) {
+		return nil, fmt.Errorf("Not authenticated.")
+	}
 
 	id, _ := params.Args["id"].(int)
 	text, _ := params.Args["text"].(string)
@@ -86,6 +139,10 @@ func EditTodo(params graphql.ResolveParams) (interface{}, error) {
 
 	if err := database.Select(&toUpdate); err != nil {
 		return nil, err
+	}
+
+	if toUpdate.UserID != user.ID {
+		return nil, fmt.Errorf("Not authrorized.")
 	}
 
 	toUpdate.Text = text
@@ -99,12 +156,21 @@ func EditTodo(params graphql.ResolveParams) (interface{}, error) {
 
 func DeleteTodo(params graphql.ResolveParams) (interface{}, error) {
 	database := params.Context.Value("database").(*pg.DB)
+	user := params.Context.Value("user").(*structs.Claims)
+
+	if !user.VerifyExpiresAt(time.Now().Unix(), true) {
+		return nil, fmt.Errorf("Not authenticated.")
+	}
 
 	id, _ := params.Args["id"].(int)
 
 	toDelete := structs.Todo{ID: id}
 	if err := database.Select(&toDelete); err != nil {
 		return nil, err
+	}
+
+	if toDelete.UserID != user.ID {
+		return nil, fmt.Errorf("Not authrorized.")
 	}
 
 	if err := database.Delete(&toDelete); err != nil {
@@ -116,11 +182,16 @@ func DeleteTodo(params graphql.ResolveParams) (interface{}, error) {
 
 func DeleteCompletedTodos(params graphql.ResolveParams) (interface{}, error) {
 	database := params.Context.Value("database").(*pg.DB)
+	user := params.Context.Value("user").(*structs.Claims)
+
+	if !user.VerifyExpiresAt(time.Now().Unix(), true) {
+		return nil, fmt.Errorf("Not authenticated.")
+	}
 
 	res, err := database.Exec(`
 		DELETE FROM todos
-		WHERE completed IS TRUE
-	`)
+		WHERE user_id = ? AND completed IS TRUE
+	`, user.ID)
 
 	if err != nil {
 		return nil, err
